requirements.txt
Django
gunicorn
django-heroku
django-allauth
django-phone-field
django-crispy-forms
django-filter
Pillow
